perf: adapt change buffer protocol to 16-bit opcodes and string indices
Update the JS↔Rust WASM shared-buffer protocol to match the changes made
in libdatadog (libdd-trace-utils change_buffer):

**Binary layout changes (JS side):**
- Buffer header: count was written as u64 (two u32 writes at offsets 0 and
  4, 8 bytes total); now a single u32 at offset 0 (4 bytes). Operations
  start at offset 4 instead of 8.
- OpCode: was written as u64 (setBigUint64, 8 bytes); now u16 (setUint16,
  2 bytes). Operation header shrinks from 16 B to 10 B.
- String indices: were written as u32 (setUint32, 4 bytes); now u16
  (setUint16, 2 bytes).
- Count flush detection: getBigUint64 + BigInt comparison → getUint32 +
  plain number comparison.

**Rust binding changes:**
- `stringTableInsertOne` / `stringTableEvict` wasm and native bindings:
  key parameter u32 → u16 (matches updated libdd-trace-utils API).
- `stringTableInsertMany`: reads key as u16 from the string table input
  buffer; added u16 FromBytes impl in utils.rs.
- Fixed stale comment: OpCode repr was documented as u64, now u16.

**Cargo.lock:** bumped libdatadog git dependency to commit c0f4b30e
(the change-buffer-perf branch commit that introduced the 16-bit encoding).

+++ crates/pipeline-native/src/lib.rs
@@ -162,14 +162,14 @@ impl NativeSpanState {
 
     /// Insert a string into the string table.
     #[napi]
-    pub fn string_table_insert_one(&self, key: u32, val: String) {
+    pub fn string_table_insert_one(&self, key: u16, val: String) {
         let mut inner = self.inner.borrow_mut();
         inner.cbs.string_table_insert_one(key, val.into());
     }
 
     /// Evict a string from the string table.
     #[napi]
-    pub fn string_table_evict(&self, key: u32) {
+    pub fn string_table_evict(&self, key: u16) {
         let mut inner = self.inner.borrow_mut();
         inner.cbs.string_table_evict_one(key);
     }
